chore: extracts tests to their own files
